requirements.txt
anyio==3.6.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
h11==0.14.0
idna==3.4
motor==3.0.0
passlib==1.7.4
pydantic==1.10.2
pymongo==4.2.0
python-dotenv==0.21.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.3.0
uvicorn==0.18.3
PyJWT==1.7.1
python-decouple==3.3
bcrypt==4.0.1
python-multipart==0.0.5
python-jose==3.3.0